Split remaining Profile CMS oversized files
