+++ scripts/debt-ratchet-baseline.json
@@ -4,7 +4,7 @@
   "counts": {
     "any_casts": 4,
     "todo_comments": 0,
-    "oversized_files": 79,
+    "oversized_files": 76,
     "bootstrap_imports": 0,
     "redux_imports": 0,
     "pages_router_files": 0
@@ -71,7 +71,6 @@
     "components/nextGen/admin/NextGenAdmin.tsx",
     "components/nft-transfer/TransferModal.tsx",
     "components/notifications/NotificationsContext.tsx",
-    "components/profile-cms-builder/ProfileCmsBuilder.tsx",
     "components/react-query-wrapper/ReactQueryWrapper.tsx",
     "components/rememes/RememePage.tsx",
     "components/user/collected/UserPageCollected.tsx",
@@ -85,8 +84,6 @@
     "helpers/Helpers.ts",
     "helpers/waves/wave-rules.helpers.ts",
     "hooks/useMemesQuickVoteQueue.ts",
-    "lib/profile-cms/builder/agent.ts",
-    "lib/profile-cms/protocol/v1/validation.ts",
     "lib/twitter/fetcher.ts",
     "services/api/wave-drops-v2-api.ts"
   ]
